Create authentication cookies file
