+++ packages/server/api/src/app/authentication/context/authentication-cookies.ts
@@ -1,0 +1,54 @@
+import {
+  AppSystemProp,
+  SharedSystemProp,
+  system,
+} from '@openops/server-shared';
+import { AuthenticationResponse } from '@openops/shared';
+import { FastifyReply } from 'fastify';
+import { jwtDecode } from 'jwt-decode';
+import { getSubDomain } from '../../helper/sub-domain';
+
+export function setAuthCookiesAndReply(
+  reply: FastifyReply,
+  response: AuthenticationResponse,
+): FastifyReply {
+  const date = jwtDecode<{ exp: number }>(response.tablesRefreshToken);
+  const cookieExpiryDate = new Date(date.exp * 1000);
+
+  return reply
+    .setCookie('jwt_token', response.tablesRefreshToken, {
+      domain: getOpenOpsSubDomain(),
+      path: '/',
+      signed: true,
+      httpOnly: false,
+      expires: cookieExpiryDate,
+    })
+    .setCookie('token', response.token, {
+      path: '/',
+      signed: true,
+      httpOnly: false,
+      expires: cookieExpiryDate,
+      sameSite: 'lax',
+    })
+    .send(response);
+}
+
+export function removeAuthCookiesAndReply(reply: FastifyReply): FastifyReply {
+  return reply
+    .clearCookie('jwt_token', {
+      domain: getOpenOpsSubDomain(),
+      path: '/',
+    })
+    .clearCookie('token', {
+      path: '/',
+    })
+    .send('Cookies removed');
+}
+
+function getOpenOpsSubDomain(): string {
+  const frontendUrl = system.getOrThrow(SharedSystemProp.FRONTEND_URL);
+
+  const tablesUrl = system.getOrThrow(AppSystemProp.OPENOPS_TABLES_PUBLIC_URL);
+
+  return getSubDomain(frontendUrl, tablesUrl);
+}
